Refactor/#72 (#73)
* refactor: page style 분리


* refactor: 자주 쓰는 type global로 분리

- props 제외 분리


* refactor: axios 통일


* refactor: list component paginator 선택 옵션, 취소, 저장 버튼 바꾸기


* fix: 마크다운에디터 옵션 수정

- 기본 에디터 마크다운으로 수정
- 마크다운 옵션에서 task, image 제거


* refactor: 모달 컴포넌트 분리


* refactor: 필요없는 props 제거
